Refactor: Split render stage classes into separate files
Improved code organization following one class per file principle.

File Structure Changes:
* PrimaryRasterizationStage.h: Only PrimaryRasterizationStage class
* SecondaryRayTracingStage.h: NEW SecondaryRayTracingStage class
* AIDenoiseStage.h: NEW AIDenoiseStage class
* UpscalingStage.h: Only UpscalingStage class
* FinalCompositeStage.h: NEW FinalCompositeStage class
* AllRenderStages.h: NEW convenience header including all stages

Benefits:
* Better code organization and maintainability
* Easier navigation and IDE support
* Follows Single Responsibility Principle at file level
* Cleaner dependencies and includes
* Convenient factory for standard pipeline creation
